fixed after review

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -12,5 +12,5 @@ public class Film {
     private String description;
     private LocalDate releaseDate;
     private int duration;
-    private Set<Long> likes = new HashSet<>();
+    private Set<Integer> likes = new HashSet<>();
 }
